Rename and sync/change namespaces
